% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_wei.R
\name{bootstrap_wei}
\alias{bootstrap_wei}
\title{Boostrap sampling -using weights-}
\usage{
bootstrap_wei(df, wei, reps = 1000)
}
\arguments{
\item{df}{a data frame.}

\item{wei}{a vector of probability weights for obtaining the elements of the vector being sampled.}

\item{reps}{yhe number of bootstrap samples.}
}
\value{
An tibble with classes bootstraps, rset, tbl_df, tbl, and data.frame. The results include a column for the data split objects and a column called id that has a character string with the resample identifier.
}
\description{
A bootstrap sample is a sample that is the same size as the original data set that is made using replacement. This results in analysis samples that have multiple replicates of some of the original rows of the data. The assessment set is defined as the rows of the original data that were not included in the bootstrap sample. This is often referred to as the "out-of-bag" -OOB- sample.
}
\examples{
df_boots <- bootstrap_wei(df=df, wei=df$wei, reps=5000);
}
