% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_wei.R
\name{sample_wei}
\alias{sample_wei}
\title{Weigthed sample index
Extracts sampling indexes from data.frame -only one sample-}
\usage{
sample_wei(df, wei)
}
\arguments{
\item{df}{a data frame, data frame extension -e.g. a tibble-, or a lazy data frame -e.g. from dbplyr or dtplyr-. See Methods, below, for more details.}

\item{wei}{a vector of probability weights for obtaining the elements of the vector being sampled.}
}
\value{
a sampled -using weights- data frame, tibble, etc.
}
\description{
Weigthed sample index
Extracts sampling indexes from data.frame -only one sample-
}
\examples{
df_sampled <- sample_wei(df=df, probs=df$wei);
}
